# web/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ayma website API",
        "description": "This API is designed to work with the backend of the site through AJAX requests from the frontend part of the application.",
        "contact": {
            "email": "mkuchirov@gmail.com"
        },
        "license": {
            "name": "API License",
            "url": "URL-to-your-api-license"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get the list of services",
                "description": "Retrieves a list of services including related service items and descriptions.",
                "operationId": "getServices",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Service"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Services not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Service": {
                "title": "Service",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier for the Service",
                        "type": "integer",
                        "format": "int64"
                    },
                    "description": {
                        "properties": {
                            "name": {
                                "description": "Name of the Service Description",
                                "type": "string"
                            },
                            "description": {
                                "description": "Detailed description of the Service",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "service_items": {
                        "description": "List of related service items",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ServiceItem"
                        }
                    }
                },
                "type": "object"
            },
            "ServiceItem": {
                "title": "Service Item",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier for the Service Item",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Name of the Service Item",
                        "type": "string"
                    },
                    "details_page_name": {
                        "description": "Name of the Service Item for details page",
                        "type": "string"
                    },
                    "description": {
                        "properties": {
                            "name": {
                                "description": "Name of the Service Item Description",
                                "type": "string"
                            },
                            "description": {
                                "description": "Detailed description of the Service Item",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}